+++ src/dw_version.h
@@ -3,7 +3,7 @@
 
 #define DW_VERSION_MAJOR "0"
 #define DW_VERSION_MINOR "2"
-#define DW_VERSION_PATCH "8"
+#define DW_VERSION_PATCH "9"
 #define deconwolf_version DW_VERSION_MAJOR "." \
     DW_VERSION_MINOR "." \
     DW_VERSION_PATCH